+++ pysui/sui/sui_grpc/pgrpc_clients.py
@@ -24,7 +24,9 @@ from pysui.sui.sui_grpc.pgrpc_async_txn import AsyncSuiTransaction
 import pysui.sui.sui_grpc.pgrpc_absreq as absreq
 from pysui.sui.sui_grpc.pgrpc_requests import GetEpoch
 
-import pysui.sui.sui_grpc.suimsgs.sui.rpc.v2beta2 as sui_prot
+
+import pysui.sui.sui_grpc.suimsgs.sui.rpc.v2 as sui_prot
+
 
 from grpclib.client import Channel
 
@@ -155,9 +157,9 @@ class SuiGrpcClient(PysuiClient):
         srv_fn: Callable[[betterproto2.Message], betterproto2.Message]
         srv_req: betterproto2.Message
         match request.service:
-            case absreq.Service.LIVEDATA:
+            case absreq.Service.STATE:
                 srv_fn, srv_req = request.to_request(
-                    stub=sui_prot.LiveDataServiceStub(self._channel)
+                    stub=sui_prot.StateServiceStub(self._channel)
                 )
             case absreq.Service.LEDGER:
                 srv_fn, srv_req = request.to_request(